% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ipw.scch}
\alias{ipw.scch}
\title{sCCH resampling}
\usage{
ipw.scch(J, D, R, V, B)
}
\arguments{
\item{J}{...}

\item{D}{...}

\item{R}{...}

\item{V}{...}

\item{B}{...}
}
\description{
sCCH resampling
}
