% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{PI.fun}
\alias{PI.fun}
\title{Calculating the baseline cumulative hazard}
\usage{
PI.fun(X, Z, beta, t = Inf, order = c(0, 1, 2), W = NULL)
}
\arguments{
\item{X}{...}

\item{Z}{...}

\item{beta}{...}

\item{t}{...}

\item{order}{...}

\item{W}{...}
}
\description{
Calculating the baseline cumulative hazard
}
