% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{survCCH-package}
\alias{survCCH-package}
\alias{survCCH}
\title{Analyzing Survival Data from Case-Cohort Studies with Perturbation Resampling}
\description{
Analyzing Survival Data from Case-Cohort Studies with Perturbation Resampling
}
\keyword{package}
