% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{survcox.scch}
\alias{survcox.scch}
\title{Main function for sCCH analysis}
\usage{
survcox.scch(N, X, D, Z, R, J, n.ptb = 1000, Z.new = NULL)
}
\arguments{
\item{N}{...}

\item{X}{...}

\item{D}{...}

\item{Z}{...}

\item{R}{...}

\item{J}{...}

\item{n.ptb}{...}

\item{Z.new}{...}
}
\description{
Main function for sCCH analysis
}
